Moved Loginmodal to Components
