+++ src/components/LoginModal/EmailSent/index.tsx
@@ -4,10 +4,10 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
 
-import { ModalType } from '../.'
+import { ModalType } from '..'
 
-import Check from '../../../../assets/images/check.png'
-import retinaCheck from '../../../../assets/images/check@2x.png'
+import Check from './../../../assets/images/check.png'
+import retinaCheck from './../../../assets/images/check@2x.png'
 
 export interface UserResetPassword {
     email: string
