+++ src/components/LoginModal/LoginOptions/index.tsx
@@ -5,8 +5,8 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton, SecondaryButton } from '@entur/button'
 import { ModalType } from '..'
 
-import sikkerhetBom from '../../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from '../../../../assets/images/sikkerhet_bom@2x.png'
+import sikkerhetBom from './../../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from './../../../assets/images/sikkerhet_bom@2x.png'
 
 interface Props {
     setModalType: Dispatch<SetStateAction<ModalType>>
