+++ src/components/LoginModal/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/import.scss';
+@import './../../assets/styles/import.scss';
 
 .login-modal.eds-modal__content {
     padding-top: 0.8rem;
